requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
google-api-python-client>=2.108.0
networkx>=3.2.0
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0
python-multipart>=0.0.6
pydantic>=2.5.0
aiohttp>=3.9.0
google-generativeai>=0.8.2


